# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Care Platform API",
        "description": "API documentation for Pet Care Platform - connecting pet owners with reliable sitters",
        "contact": {
            "email": "admin@petcare.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sitter-services": {
            "get": {
                "tags": [
                    "Sitter Services"
                ],
                "summary": "Get all sitter services (public)",
                "description": "Returns list of sitter services with optional filtering - no authentication required",
                "operationId": "getSitterServicesList",
                "parameters": [
                    {
                        "name": "service_type_id",
                        "in": "query",
                        "description": "Filter by service type ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sitter services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "hourly_rate": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "service_type_id": {
                                                "type": "integer"
                                            },
                                            "sitter_id": {
                                                "type": "integer"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "service_type": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameters provided"
                    },
                    "404": {
                        "description": "Not Found - The requested resource was not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation failed for the input data"
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred on the server"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sitter Services"
                ],
                "summary": "Create a new sitter service",
                "operationId": "d5eb8601dfee00d018433abe35eb47ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "hourly_rate",
                                    "service_type_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dog Walking Service"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Professional dog walking services"
                                    },
                                    "hourly_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25
                                    },
                                    "service_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sitter_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sitter service created successfully"
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameters provided"
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid authentication token"
                    },
                    "403": {
                        "description": "Forbidden - You do not have access to this resource"
                    },
                    "404": {
                        "description": "Not Found - The requested resource was not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation failed for the input data"
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred on the server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sitter-services/{id}": {
            "get": {
                "tags": [
                    "Sitter Services"
                ],
                "summary": "Get a specific sitter service (public)",
                "description": "Get specific sitter service details - no authentication required",
                "operationId": "e069b2e1db039df34a81837da028a1eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sitter service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sitter service details"
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameters provided"
                    },
                    "404": {
                        "description": "Not Found - Sitter service not found"
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred on the server"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sitter Services"
                ],
                "summary": "Update a sitter service",
                "operationId": "441d3554aec7019144d5e66bbbd56c3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sitter service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "hourly_rate",
                                    "service_type_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dog Walking Service"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Professional dog walking services"
                                    },
                                    "hourly_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25
                                    },
                                    "service_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sitter_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sitter service updated successfully"
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameters provided"
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid authentication token"
                    },
                    "403": {
                        "description": "Forbidden - You do not have access to this resource"
                    },
                    "404": {
                        "description": "Not Found - Sitter service not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation failed for the input data"
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred on the server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sitter Services"
                ],
                "summary": "Delete a sitter service",
                "operationId": "adf288631c77ad3c2e480b7ed4810f95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sitter service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sitter service deleted successfully"
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameters provided"
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid authentication token"
                    },
                    "403": {
                        "description": "Forbidden - You do not have access to this resource"
                    },
                    "404": {
                        "description": "Not Found - Sitter service not found"
                    },
                    "500": {
                        "description": "Internal Server Error - An unexpected error occurred on the server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "scheme": "bearer"
            },
            "0": []
        }
    },
...
}